U2 - Changed function name to match coding standards

+++ Pages/ProductPage.xaml.cs
@@ -7,18 +7,18 @@ namespace ENF_Dist_Test.Pages{
 
         public ProductPage(){
             InitializeComponent();
-            updateButtons(null, null);
-            updateTable();
+            UpdateButtons(null, null);
+            UpdateTable();
         }
-        private void updateTable() {
+        private void UpdateTable() {
             DataGrid.ItemsSource = Database.Instance.GetAllProducts();
         }
-        public void updateButtons(object sender, RoutedEventArgs e) {
-            UpdateBtn.IsEnabled = hasSelected;
-            DeleteBtn.IsEnabled = hasSelected;
+        public void UpdateButtons(object? sender, RoutedEventArgs? e) {
+            UpdateBtn.IsEnabled = HasSelected;
+            DeleteBtn.IsEnabled = HasSelected;
         }
 
-        public bool hasSelected {
+        public bool HasSelected {
             get { return DataGrid.SelectedItem != null; }
         }
         private void NavBack(object sender, RoutedEventArgs e){
@@ -26,23 +26,28 @@ namespace ENF_Dist_Test.Pages{
         }
 
         private void Add(object sender, RoutedEventArgs e) {
-            ProductEdit productEdit = new ProductEdit(new() { ProductId = Database.Instance.GetNextID("Products")}, false);
-            productEdit.ShowDialog();
-            if (!productEdit.AddCancel) {
-                Database.Instance.InsertLocation(productEdit.product.Location);
-                Database.Instance.InsertProduct(productEdit.product);
-                updateTable();
-            }
-        }
-        private void Update(object sender, RoutedEventArgs e) {
-            ProductEdit productEdit = new ProductEdit((Product)DataGrid.SelectedItem, true);
-            string prevLocation = productEdit.product.Location.LocationId;
+            ProductEdit productEdit = new(new() { ProductId = Database.Instance.GetNextID("Products")}, false);
             productEdit.ShowDialog();
             if (!productEdit.AddCancel) {
                 Database.Instance.InsertLocation(productEdit.Product.Location);
-                Database.Instance.UpdateProduct(productEdit.product, productEdit.product.ProductId);
-                Database.Instance.DeleteLocation(prevLocation);
-                updateTable();
+                Database.Instance.InsertProduct(productEdit.Product);
+                UpdateTable();
+            }
+        }
+        private void Update(object sender, RoutedEventArgs e) {
+            ProductEdit productEdit = new((Product)DataGrid.SelectedItem, true);
+            string prevLocation = productEdit.Product.Location.LocationId;
+            productEdit.ShowDialog();
+            if (!productEdit.AddCancel) {
+                if(productEdit.oldLoc != productEdit.Product.Location.LocationId) {
+                    Database.Instance.InsertLocation(productEdit.Product.Location);
+                }
+                Database.Instance.UpdateProduct(productEdit.Product, productEdit.Product.ProductId);
+
+                if (productEdit.oldLoc != productEdit.Product.Location.LocationId) {
+                    Database.Instance.DeleteLocation(prevLocation);
+                }
+                UpdateTable();
             }
         }
         private void Delete(object sender, RoutedEventArgs e) {
@@ -50,7 +55,7 @@ namespace ENF_Dist_Test.Pages{
             if (MessageBox.Show($"Are you sure you want to delete {product}?", "Delete", MessageBoxButton.YesNo, MessageBoxImage.Warning) == MessageBoxResult.Yes) {
                 Database.Instance.DeleteProduct(product.ProductId);
                 Database.Instance.DeleteLocation(product.Location.LocationId);
-                updateTable();
+                UpdateTable();
             }
         }
     }
